feat(client): separates validation schemas into their own files with strong typing
