+++ src/client/src/content/forms/common/validation/typeReferenceListSchema.ts
@@ -1,0 +1,8 @@
+import * as yup from "yup";
+
+export const typeReferenceListSchema = (errorText: string) =>
+  yup.array().of(
+    yup.object().shape({
+      name: yup.string().required(errorText),
+    })
+  );
