+++ src/client/src/content/forms/common/validation/valueObjectListSchema.ts
@@ -1,0 +1,8 @@
+import * as yup from "yup";
+
+export const valueObjectListSchema = (errorText: string) =>
+  yup.array().of(
+    yup.object().shape({
+      value: yup.string().required(errorText),
+    })
+  );
